pre-final fix bugs, add between rating

+++ src/main/java/workplace/service/VoiceService.java
@@ -3,6 +3,7 @@ package workplace.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 import workplace.model.Voice;
@@ -11,7 +12,6 @@ import workplace.repository.user.DataJpaUserRepository;
 import workplace.repository.voice.DataJpaVoiceRepository;
 import workplace.to.VoiceTo;
 import workplace.util.Util;
-import workplace.util.VoiceUtil;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -35,31 +35,31 @@ public class VoiceService {
     @CacheEvict(value = "voices", allEntries = true)
     public Voice create(Voice voice, int userId){
         Assert.notNull(voice, " voice must be not null ");
-        /*if (Util.checkVoteTime()) {*/
+        if (Util.checkVoteTime()) {
             return voiceRepository.save(voice, userId);
-        /*}
+        }
         else{
             return null;
-        }*/
+        }
     }
 
     @CacheEvict(value = "voices", allEntries = true)
     public Voice create(VoiceTo voiceTo, int userId){
         Assert.notNull(voiceTo, " voice must be not null ");
         Voice voice = new Voice(null, voiceTo.getDate(), restaurantRepository.get(voiceTo.getRestaurantId()), userRepository.get(voiceTo.getUserId()));
-        /*if (Util.checkVoteTime()) {*/
+        if (Util.checkVoteTime()) {
             return voiceRepository.save(voice, userId);
-        /*}
+        }
         else{
             return null;
-        }*/
+        }
     }
 
     @CacheEvict(value = "voices", allEntries = true)
     public void delete(int id, int userId){
-        /*if(Util.checkVoteTime()) {*/
+        if(Util.checkVoteTime()) {
             checkNotFoundWithId(voiceRepository.delete(id, userId), id);
-        /*}*/
+        }
     }
 
     public Voice get(int id, int userId){
@@ -68,10 +68,10 @@ public class VoiceService {
 
     @CacheEvict(value = "voices", allEntries = true)
     public void update(Voice voice, int userId){
-        /*Assert.notNull(voice, " voice must be not null ");
-        if(Util.checkVoteTime()) {*/
+        Assert.notNull(voice, " voice must be not null ");
+        if(Util.checkVoteTime()) {
         checkNotFoundWithId(voiceRepository.save(voice, userId), voice.getId());
-        /*}*/
+        }
     }
 
     @CacheEvict(value = "voices", allEntries = true)
@@ -79,9 +79,9 @@ public class VoiceService {
         Assert.notNull(voiceTo, " voiceTo must be not null ");
         Voice voice = get(voiceTo.getId(), userId);
         voice.setRestaurant(restaurantRepository.get(voiceTo.getRestaurantId()));
-        /*if(Util.checkVoteTime()) {*/
+        if(Util.checkVoteTime()) {
             checkNotFoundWithId(voiceRepository.save(voice, userId), voice.getId());
-        /*}*/
+        }
 
     }
 
@@ -92,5 +92,9 @@ public class VoiceService {
 
     public int getRating(int restaurantId, LocalDate date){
         return voiceRepository.getRateByRestaurant(restaurantId, date);
+    }
+
+    public int getBetweenDates(int restaurantId, @Nullable LocalDate startDate, @Nullable LocalDate endDate){
+        return voiceRepository.getRatingBetween(restaurantId, Util.adjustStartDate(startDate), Util.adjustEndDate(endDate));
     }
 }
